requirements.txt
PyQt6>=6.4.0
opencv-python>=4.6.0
numpy>=1.23.0
pycryptodome>=3.15.0
pytesseract
pillow